Refactored Map code by removing comments, code, and moving them
