+++ frontend/components/map-components/styles/MapIcons.js
